Github issue templates will now auto-assign labels
We have updated the Github issue templates to auto-assign
the appropriate label:

(1) "Ask a question" auto-assigns the "question" label
(2) "Report a bug/issue" auto-assigns the "bug" label
(3) "Request a feature" auto-assigns the "feature" label

We have also added the "Start a discussion" issue template, which
is intended for longer discussions (about open or unresolved topics).
This will auto-assign the "discussion" label.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>